+++ app/src/main/java/com/example/text_it/Activity/LandingUser.kt
@@ -4,6 +4,7 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
 import android.widget.LinearLayout
+import android.widget.TextView
 import android.widget.Toast
 import com.example.text_it.R
 
@@ -12,14 +13,8 @@ class LandingUser : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_landing_user)
 
-        if (checkSelfPermission(android.Manifest.permission.CAMERA) != android.content.pm.PackageManager.PERMISSION_GRANTED) {
-            requestPermissions(arrayOf(
-                android.Manifest.permission.CAMERA
-            ), 1)
-        }
-
         val signupButton: Button = findViewById(R.id.signupButton)
-        val loginButton: LinearLayout = findViewById(R.id.loginLayout)
+        val loginButton: TextView = findViewById(R.id.textViewLogin)
 
         signupButton.setOnClickListener {
             startActivity(
@@ -38,17 +33,5 @@ class LandingUser : AppCompatActivity() {
         }
 
 
-    }
-
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<out String>,
-        grantResults: IntArray
-    ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        if(grantResults[0] != android.content.pm.PackageManager.PERMISSION_GRANTED){
-            Toast.makeText(this, "Permissions not granted", Toast.LENGTH_SHORT).show()
-            finish()
-        }
     }
 }
