+++ app/src/main/res/layout/activity_landing_user.xml
@@ -67,30 +67,29 @@
     </LinearLayout>
 
     <LinearLayout
+        android:id="@+id/orLayout"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_below="@+id/socialButtons"
-        android:id="@+id/orLayout"
-        android:orientation="horizontal"
-        android:gravity="center_vertical"
         android:layout_centerHorizontal="true"
-        android:layout_marginTop="30dp">
+        android:layout_marginTop="30dp"
+        android:gravity="center_vertical"
+        android:orientation="horizontal">
 
         <View
             android:layout_width="135dp"
             android:layout_height="1dp"
-            android:background="#33ffffff"
-            />
+            android:background="#33ffffff" />
 
         <TextView
             android:id="@+id/orText"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_centerInParent="true"
             android:layout_marginLeft="15dp"
             android:layout_marginRight="15dp"
             android:fontFamily="@font/poppins_bold"
             android:text="OR"
-            android:layout_centerInParent="true"
             android:textColor="#ffffff"
             android:textSize="15sp" />
 
@@ -102,6 +101,7 @@
     </LinearLayout>
 
     <Button
+        android:id="@+id/signupButton"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_below="@+id/orLayout"
@@ -110,36 +110,37 @@
         android:fontFamily="@font/poppins"
         android:text="Sign up with email"
         android:textColor="#ffffff"
-        android:textSize="15sp"
-        android:id="@+id/signupButton"/>
+        android:textSize="15sp" />
 
     <LinearLayout
+        android:id="@+id/loginLayout"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:orientation="horizontal"
         android:layout_below="@+id/signupButton"
-        android:layout_marginTop="25dp"
         android:layout_centerInParent="true"
+        android:layout_marginTop="25dp"
         android:clickable="true"
-        android:id="@+id/loginLayout">
+        android:orientation="horizontal">
+
         <TextView
+            android:id="@+id/textViewAlreadyHaveAnAccount"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_marginRight="5dp"
             android:fontFamily="@font/poppins"
             android:text="Already have an account?"
             android:textColor="#ffffff"
-            android:textSize="15sp"
-            android:layout_marginRight="5dp"
-            android:id="@+id/textViewAlreadyHaveAnAccount"/>
+            android:textSize="15sp" />
+
         <TextView
+            android:id="@+id/textViewLogin"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_below="@+id/textViewAlreadyHaveAnAccount"
-            android:fontFamily="@font/poppins_bold"
             android:layout_marginLeft="5dp"
+            android:fontFamily="@font/poppins_bold"
             android:text="Log in"
             android:textColor="#ffffff"
-            android:textSize="15sp"
-            android:id="@+id/textViewLogin"/>
+            android:textSize="15sp" />
     </LinearLayout>
 </RelativeLayout>
